feat: refactor the tool, use one file only
