+++ src/strands_tools/apify.py
@@ -1,11 +1,17 @@
-"""Shared base for Apify platform tools.
+"""Apify platform tools for Strands Agents.
 
-This module provides the shared infrastructure used by all Apify tool modules
-(e.g. apify_core, apify_social). It contains the API client, error handling,
-response helpers, and constants. It does NOT contain any @tool functions itself.
+This module provides web scraping, data extraction, and automation capabilities
+using the Apify platform. It lets you run any Actor, task, fetch dataset
+results, and scrape individual URLs.
 
-Tool modules import from here:
-    from strands_tools.apify import ApifyToolClient, _check_dependency, ...
+Available Tools:
+---------------
+- apify_run_actor: Run any Apify Actor with custom input
+- apify_get_dataset_items: Fetch items from an Apify dataset with pagination
+- apify_run_actor_and_get_dataset: Run an Actor and fetch results in one step
+- apify_run_task: Run a saved Actor task with optional input overrides
+- apify_run_task_and_get_dataset: Run a task and fetch results in one step
+- apify_scrape_url: Scrape a single URL and return content as Markdown
 
 Setup Requirements:
 ------------------
@@ -14,8 +20,41 @@ Setup Requirements:
 3. Install the optional dependency: pip install strands-agents-tools[apify]
 4. Set the environment variable:
    APIFY_API_TOKEN=your_api_token_here
+
+Usage Examples:
+--------------
+Register all core tools at once via the preset list:
+
+```python
+from strands import Agent
+from strands_tools.apify import APIFY_CORE_TOOLS
+
+agent = Agent(tools=APIFY_CORE_TOOLS)
+```
+
+Or pick individual tools for a smaller LLM tool surface:
+
+```python
+from strands import Agent
+from strands_tools import apify
+
+agent = Agent(tools=[
+    apify.apify_scrape_url,
+    apify.apify_run_actor,
+])
+
+# Scrape a single URL
+content = agent.tool.apify_scrape_url(url="https://example.com")
+
+# Run an Actor
+result = agent.tool.apify_run_actor(
+    actor_id="apify/website-content-crawler",
+    run_input={"startUrls": [{"url": "https://example.com"}]},
+)
+```
 """
 
+import json
 import logging
 import os
 from typing import Any, Dict, List, Optional
@@ -23,6 +62,7 @@ from urllib.parse import urlparse
 
 from rich.panel import Panel
 from rich.text import Text
+from strands import tool
 
 from strands_tools.utils import console_util
 
@@ -311,3 +351,300 @@ class ApifyToolClient:
             raise RuntimeError(f"No content returned for URL: {url}")
 
         return str(items[0].get("markdown") or items[0].get("text", ""))
+
+
+# --- Tool functions ---
+
+
+@tool
+def apify_run_actor(
+    actor_id: str,
+    run_input: Optional[Dict[str, Any]] = None,
+    timeout_secs: int = DEFAULT_TIMEOUT_SECS,
+    memory_mbytes: Optional[int] = None,
+    build: Optional[str] = None,
+) -> Dict[str, Any]:
+    """Run any Apify Actor and return the run metadata as JSON.
+
+    Executes the Actor synchronously - blocks until the Actor run finishes or the timeout
+    is reached. Use this when you need to run a specific Actor and then inspect or process
+    the results separately.
+
+    Common Actors:
+    - "apify/website-content-crawler" - scrape websites and extract content
+    - "apify/web-scraper" - general-purpose web scraper
+    - "apify/google-search-scraper" - scrape Google search results
+
+    Args:
+        actor_id: Actor identifier, e.g. "apify/website-content-crawler" or "username/actor-name".
+        run_input: JSON-serializable input for the Actor. Each Actor defines its own input schema.
+        timeout_secs: Maximum time in seconds to wait for the Actor run to finish. Defaults to 300.
+        memory_mbytes: Memory allocation in MB for the Actor run. Uses Actor default if not set.
+        build: Actor build tag or number to run a specific version. Uses latest build if not set.
+
+    Returns:
+        Dict with status and content containing run metadata: run_id, status, dataset_id,
+        started_at, finished_at.
+    """
+    try:
+        _check_dependency()
+        client = ApifyToolClient()
+        result = client.run_actor(
+            actor_id=actor_id,
+            run_input=run_input,
+            timeout_secs=timeout_secs,
+            memory_mbytes=memory_mbytes,
+            build=build,
+        )
+        return _success_result(
+            text=json.dumps(result, indent=2, default=str),
+            panel_body=(
+                f"[green]Actor run completed[/green]\n"
+                f"Actor: {actor_id}\n"
+                f"Run ID: {result['run_id']}\n"
+                f"Status: {result['status']}\n"
+                f"Dataset ID: {result['dataset_id']}"
+            ),
+            panel_title="Apify: Run Actor",
+        )
+    except Exception as e:
+        return _error_result(e, "apify_run_actor")
+
+
+@tool
+def apify_get_dataset_items(
+    dataset_id: str,
+    limit: int = DEFAULT_DATASET_ITEMS_LIMIT,
+    offset: int = 0,
+) -> Dict[str, Any]:
+    """Fetch items from an existing Apify dataset and return them as JSON.
+
+    Use this after running an Actor to retrieve the structured results from its
+    default dataset, or to access any dataset by ID.
+
+    Args:
+        dataset_id: The Apify dataset ID to fetch items from.
+        limit: Maximum number of items to return. Defaults to 100.
+        offset: Number of items to skip for pagination. Defaults to 0.
+
+    Returns:
+        Dict with status and content containing an array of dataset items.
+    """
+    try:
+        _check_dependency()
+        client = ApifyToolClient()
+        items = client.get_dataset_items(dataset_id=dataset_id, limit=limit, offset=offset)
+        return _success_result(
+            text=json.dumps(items, indent=2, default=str),
+            panel_body=(
+                f"[green]Dataset items retrieved[/green]\nDataset ID: {dataset_id}\nItems returned: {len(items)}"
+            ),
+            panel_title="Apify: Dataset Items",
+        )
+    except Exception as e:
+        return _error_result(e, "apify_get_dataset_items")
+
+
+@tool
+def apify_run_actor_and_get_dataset(
+    actor_id: str,
+    run_input: Optional[Dict[str, Any]] = None,
+    timeout_secs: int = DEFAULT_TIMEOUT_SECS,
+    memory_mbytes: Optional[int] = None,
+    build: Optional[str] = None,
+    dataset_items_limit: int = DEFAULT_DATASET_ITEMS_LIMIT,
+    dataset_items_offset: int = 0,
+) -> Dict[str, Any]:
+    """Run an Apify Actor and fetch its dataset results in one step.
+
+    Convenience tool that combines running an Actor and fetching its default
+    dataset items into a single call. Use this when you want both the run metadata and the
+    result data without making two separate tool calls.
+
+    Args:
+        actor_id: Actor identifier, e.g. "apify/website-content-crawler" or "username/actor-name".
+        run_input: JSON-serializable input for the Actor.
+        timeout_secs: Maximum time in seconds to wait for the Actor run to finish. Defaults to 300.
+        memory_mbytes: Memory allocation in MB for the Actor run.
+        build: Actor build tag or number to run a specific version. Uses latest build if not set.
+        dataset_items_limit: Maximum number of dataset items to return. Defaults to 100.
+        dataset_items_offset: Number of dataset items to skip for pagination. Defaults to 0.
+
+    Returns:
+        Dict with status and content containing run metadata (run_id, status, dataset_id,
+        started_at, finished_at) plus an "items" array containing the dataset results.
+    """
+    try:
+        _check_dependency()
+        client = ApifyToolClient()
+        result = client.run_actor_and_get_dataset(
+            actor_id=actor_id,
+            run_input=run_input,
+            timeout_secs=timeout_secs,
+            memory_mbytes=memory_mbytes,
+            build=build,
+            dataset_items_limit=dataset_items_limit,
+            dataset_items_offset=dataset_items_offset,
+        )
+        return _success_result(
+            text=json.dumps(result, indent=2, default=str),
+            panel_body=(
+                f"[green]Actor run completed with dataset[/green]\n"
+                f"Actor: {actor_id}\n"
+                f"Run ID: {result['run_id']}\n"
+                f"Status: {result['status']}\n"
+                f"Dataset ID: {result['dataset_id']}\n"
+                f"Items returned: {len(result['items'])}"
+            ),
+            panel_title="Apify: Run Actor + Dataset",
+        )
+    except Exception as e:
+        return _error_result(e, "apify_run_actor_and_get_dataset")
+
+
+@tool
+def apify_run_task(
+    task_id: str,
+    task_input: Optional[Dict[str, Any]] = None,
+    timeout_secs: int = DEFAULT_TIMEOUT_SECS,
+    memory_mbytes: Optional[int] = None,
+) -> Dict[str, Any]:
+    """Run an Apify task and return the run metadata as JSON.
+
+    Tasks are saved Actor configurations with preset inputs. Use this when a task
+    has already been configured in Apify Console, so you don't need to specify
+    the full Actor input every time.
+
+    Args:
+        task_id: Task identifier, e.g. "user/my-task" or a task ID string.
+        task_input: Optional JSON-serializable input to override the task's default input.
+        timeout_secs: Maximum time in seconds to wait for the task run to finish. Defaults to 300.
+        memory_mbytes: Memory allocation in MB for the task run. Uses task default if not set.
+
+    Returns:
+        Dict with status and content containing run metadata: run_id, status, dataset_id,
+        started_at, finished_at.
+    """
+    try:
+        _check_dependency()
+        client = ApifyToolClient()
+        result = client.run_task(
+            task_id=task_id,
+            task_input=task_input,
+            timeout_secs=timeout_secs,
+            memory_mbytes=memory_mbytes,
+        )
+        return _success_result(
+            text=json.dumps(result, indent=2, default=str),
+            panel_body=(
+                f"[green]Task run completed[/green]\n"
+                f"Task: {task_id}\n"
+                f"Run ID: {result['run_id']}\n"
+                f"Status: {result['status']}\n"
+                f"Dataset ID: {result['dataset_id']}"
+            ),
+            panel_title="Apify: Run Task",
+        )
+    except Exception as e:
+        return _error_result(e, "apify_run_task")
+
+
+@tool
+def apify_run_task_and_get_dataset(
+    task_id: str,
+    task_input: Optional[Dict[str, Any]] = None,
+    timeout_secs: int = DEFAULT_TIMEOUT_SECS,
+    memory_mbytes: Optional[int] = None,
+    dataset_items_limit: int = DEFAULT_DATASET_ITEMS_LIMIT,
+    dataset_items_offset: int = 0,
+) -> Dict[str, Any]:
+    """Run an Apify task and fetch its dataset results in one step.
+
+    Convenience tool that combines running a task and fetching its default
+    dataset items into a single call. Use this when you want both the run metadata and the
+    result data without making two separate tool calls.
+
+    Args:
+        task_id: Task identifier, e.g. "user/my-task" or a task ID string.
+        task_input: Optional JSON-serializable input to override the task's default input.
+        timeout_secs: Maximum time in seconds to wait for the task run to finish. Defaults to 300.
+        memory_mbytes: Memory allocation in MB for the task run.
+        dataset_items_limit: Maximum number of dataset items to return. Defaults to 100.
+        dataset_items_offset: Number of dataset items to skip for pagination. Defaults to 0.
+
+    Returns:
+        Dict with status and content containing run metadata (run_id, status, dataset_id,
+        started_at, finished_at) plus an "items" array containing the dataset results.
+    """
+    try:
+        _check_dependency()
+        client = ApifyToolClient()
+        result = client.run_task_and_get_dataset(
+            task_id=task_id,
+            task_input=task_input,
+            timeout_secs=timeout_secs,
+            memory_mbytes=memory_mbytes,
+            dataset_items_limit=dataset_items_limit,
+            dataset_items_offset=dataset_items_offset,
+        )
+        return _success_result(
+            text=json.dumps(result, indent=2, default=str),
+            panel_body=(
+                f"[green]Task run completed with dataset[/green]\n"
+                f"Task: {task_id}\n"
+                f"Run ID: {result['run_id']}\n"
+                f"Status: {result['status']}\n"
+                f"Dataset ID: {result['dataset_id']}\n"
+                f"Items returned: {len(result['items'])}"
+            ),
+            panel_title="Apify: Run Task + Dataset",
+        )
+    except Exception as e:
+        return _error_result(e, "apify_run_task_and_get_dataset")
+
+
+@tool
+def apify_scrape_url(
+    url: str,
+    timeout_secs: int = DEFAULT_SCRAPE_TIMEOUT_SECS,
+    crawler_type: str = "cheerio",
+) -> Dict[str, Any]:
+    """Scrape a single URL and return its content as markdown.
+
+    Uses the Website Content Crawler Actor under the hood, pre-configured for
+    fast single-page scraping. This is the simplest way to extract readable content
+    from any web page.
+
+    Args:
+        url: The URL to scrape, e.g. "https://example.com".
+        timeout_secs: Maximum time in seconds to wait for scraping to finish. Defaults to 120.
+        crawler_type: Crawler engine to use. One of "cheerio" (fastest, no JS rendering,
+            default), "playwright:adaptive" (fast, renders JS if present), or
+            "playwright:firefox" (reliable, renders JS, best at avoiding blocking but slower).
+
+    Returns:
+        Dict with status and content containing the markdown content of the scraped page.
+    """
+    try:
+        _check_dependency()
+        client = ApifyToolClient()
+        content = client.scrape_url(url=url, timeout_secs=timeout_secs, crawler_type=crawler_type)
+        return _success_result(
+            text=content,
+            panel_body=(
+                f"[green]URL scraped successfully[/green]\nURL: {url}\nContent length: {len(content)} characters"
+            ),
+            panel_title="Apify: Scrape URL",
+        )
+    except Exception as e:
+        return _error_result(e, "apify_scrape_url")
+
+
+APIFY_CORE_TOOLS = [
+    apify_run_actor,
+    apify_get_dataset_items,
+    apify_run_actor_and_get_dataset,
+    apify_run_task,
+    apify_run_task_and_get_dataset,
+    apify_scrape_url,
+]
